requirements.txt
requests
beautifulsoup4
lxml
selenium
undetected-chromedriver
python-dotenv
xvfbwrapper
boto3
